+++ intact-search-interactions-service/src/main/java/uk/ac/ebi/intact/search/interactions/repository/CustomizedChildInteractorRepository.java
@@ -41,7 +41,7 @@ public interface CustomizedChildInteractorRepository {
                                                           double minMiScore,
                                                           double maxMiScore,
                                                           boolean interSpecies,
-                                                          Set<Integer> binaryInteractionIdFilter,
+                                                          Set<Long> binaryInteractionIdFilter,
                                                           Set<String> interactorAcFilter,
                                                           Sort sort, Pageable pageable);
 
@@ -70,6 +70,6 @@ public interface CustomizedChildInteractorRepository {
                                double minMiScore,
                                double maxMiScore,
                                boolean interSpecies,
-                               Set<Integer> binaryInteractionIdFilter,
+                               Set<Long> binaryInteractionIdFilter,
                                Set<String> interactorAcFilter);
 }
