+++ intact-search-interactions-service/src/main/java/uk/ac/ebi/intact/search/interactions/service/ChildInteractorSearchService.java
@@ -32,7 +32,7 @@ public class ChildInteractorSearchService {
                                                                      double minMiScore,
                                                                      double maxMiScore,
                                                                      boolean interSpecies,
-                                                                     Set<Integer> binaryInteractionIdFilter,
+                                                                     Set<Long> binaryInteractionIdFilter,
                                                                      Set<String> interactorAcFilter,
                                                                      int page,
                                                                      int pageSize) {
@@ -52,7 +52,7 @@ public class ChildInteractorSearchService {
                                           double minMiScore,
                                           double maxMiScore,
                                           boolean interSpecies,
-                                          Set<Integer> binaryInteractionIdFilter,
+                                          Set<Long> binaryInteractionIdFilter,
                                           Set<String> interactorAcFilter) {
         return childInteractorRepository.countChildInteractors(query, batchSearch, interactorSpeciesFilter, interactorTypeFilter, interactionDetectionMethodFilter,
                 interactionTypeFilter, interactionHostOrganismFilter, isNegativeFilter, minMiScore, maxMiScore, interSpecies, binaryInteractionIdFilter, interactorAcFilter);
